Remove parentheses from candidate title when match initially fails
Track titles may have parentheses in them that note that the song features another artist (eg. (feat. artist)).

This means that, if a track title has this note on Spotify, but not Deezer (or vice versa), the "is title matching" check fails.

This is fixed by removing these notes from the title when comparing.